Rename tickets cubit folder
